loader/ncch: split NCCH parsing into its own file
